01-docker-terraform/pipeline/requirements.txt
pandas
pyarrow
sqlalchemy
psycopg2-binary
python-dotenv
jupyter